separated out property basics and details
