+++ app/components/Listing/Container.vue
@@ -1,5 +1,5 @@
 <template>
-  <section class="rounded border-1 border-gray-200 bg-gray-50 p-8 w-3/4 lg:w-1/2 sm:w-[90%]">
+  <section class="rounded border-1 border-gray-200 bg-gray-50 p-8 w-3/4 xl:w-2/3 2xl:w-1/2 sm:w-[90%]">
     <slot />
   </section>
 </template>
