+++ app/composables/useZodSchema.ts
@@ -1,0 +1,3 @@
+import * as z from "zod/v4/mini";
+
+export default function () {}
